Compliant Solution

+++ R02_XP00_J.java
@@ -2,6 +2,12 @@ public void deleteFile(){
  
   File someFile = new File("someFileName.txt");
   // Do something with someFile
+
+ if (!someFile.delete()) {
+    // Handle failure to delete the file
+  }
+
   someFile.delete();
  
+
 }
